+++ src/stories/components/Bullet/data.ts
@@ -12,9 +12,9 @@
 
 export const basicBulletData = [
     {"graphLabel": "Facebook", "currentAmount": 390, "target": 50},
-    {"graphLabel": "X", "currentAmount": 400, "target": 200},
-    {"graphLabel": "LinkedIn", "currentAmount": 400, "target": 200},
-    {"graphLabel": "Snapchat", "currentAmount": 400, "target": 200},
-    {"graphLabel": "Instagram", "currentAmount": 400, "target": 200},
-    {"graphLabel": "bereal", "currentAmount": 400, "target": 200},
+    {"graphLabel": "X", "currentAmount": 275, "target": 200},
+    {"graphLabel": "LinkedIn", "currentAmount": 200, "target": 450},
+    {"graphLabel": "Snapchat", "currentAmount": 250, "target": 200},
+    {"graphLabel": "Instagram", "currentAmount": 300, "target": 250},
+    {"graphLabel": "WhatsApp", "currentAmount": 375, "target": 300},
 ];
